Use ta_log_arg instead of uint32_t
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@20254 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/loggerta/logger_ta_internal.h
@@ -200,7 +200,7 @@ typedef struct md_list {
 } md_list;
 
 /** Type of argument native for a stack */
-typedef uint32_t ta_log_arg;
+typedef long ta_log_arg;
 
 /*
  * It defines the structure of the message header.
